Move exposure computation to exposure.py
