+++ gammapy/cube/__init__.py
@@ -4,6 +4,7 @@ Sky cubes (3-dimensional: energy, lon, lat).
 """
 from .models import *
 from .counts import *
+from .exposure import *
 from .fit import *
 from .new import *
 from .psf_kernel import *
