+++ gammapy/cube/counts.py
@@ -1,6 +1,4 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
-"""Functions to perform basic functions for map and cube analysis.
-"""
 from __future__ import absolute_import, division, print_function, unicode_literals
 from ..maps import WcsNDMap
