+++ gammapy/cube/tests/test_new.py
@@ -1,16 +1,15 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 from __future__ import absolute_import, division, print_function, unicode_literals
 import pytest
-import numpy as np
 from numpy.testing import assert_allclose
 import astropy.units as u
 from astropy.units import Quantity
 from astropy.coordinates import Angle, SkyCoord
 from ...utils.testing import assert_quantity_allclose
 from ...utils.testing import requires_data
-from ...irf import EffectiveAreaTable2D, Background3D
+from ...irf import Background3D
 from ...maps import WcsNDMap, WcsGeom, MapAxis
-from ..new import make_map_exposure_true_energy, make_map_background_irf, MapMaker
+from ..new import make_map_background_irf, MapMaker
 from ...data import DataStore
 
 pytest.importorskip('scipy')
@@ -23,12 +22,6 @@ def bkg_3d():
 
 
 @pytest.fixture(scope='session')
-def aeff():
-    filename = '$GAMMAPY_EXTRA/datasets/hess-crab4-hd-hap-prod2/run023400-023599/run023523/hess_aeff_2d_023523.fits.gz'
-    return EffectiveAreaTable2D.read(filename, hdu='AEFF_2D')
-
-
-@pytest.fixture(scope='session')
 def counts_cube():
     import os
     filename = os.path.join(
@@ -36,21 +29,6 @@ def counts_cube():
         'datasets/hess-crab4-hd-hap-prod2/hess_events_simulated_023523_cntcube.fits'
     )
     return WcsNDMap.read(filename)
-
-
-@requires_data('gammapy-extra')
-def test_make_map_exposure_true_energy(aeff, counts_cube):
-    pointing = SkyCoord(83.633, 21.514, unit='deg')
-    livetime = Quantity(1581.17, 's')
-    offset_max = Angle(2.2, 'deg')
-
-    m = make_map_exposure_true_energy(
-        pointing, livetime, aeff, counts_cube.geom, offset_max,
-    )
-
-    assert m.data.shape == (15, 120, 200)
-    assert m.unit == 'm2 s'
-    assert_quantity_allclose(np.nanmax(m.data), 4.7e8, rtol=100)
 
 
 @requires_data('gammapy-extra')
@@ -93,4 +71,3 @@ def test_MapMaker(mode, expected):
     maps = maker.run(obslist)
     assert maps['exposure_map'].unit == "m2 s"
     assert_quantity_allclose(maps['counts_map'].data.sum(), expected)
-
